Task token validation: validates task/pod.
